Works on Linux. Does it still work on Windows?
git-svn-id: https://enigma-dev.svn.sourceforge.net/svnroot/enigma-dev/trunk@136 726d713c-bd1a-46cb-9ef0-1b6e075168bb

+++ ENIGMAsystem/SHELL/Platforms/windows/WINDOWSStd.h
@@ -31,7 +31,7 @@ int screen_refresh();
 
 int sleep(int millis);
 void enigma_catchmouse_backend(bool x);
-int show_error(std::string errortext, double fatal);
+int show_error(std::string errortext, int fatal);
 #define enigmacatchmouse() enigma_catchmouse_backend(enigma::mousestatus[0]==1 && enigma::last_mousestatus[0]==1)
