+++ ENIGMAsystem/SHELL/Platforms/windows/WINDOWSshow_error.cpp
@@ -45,7 +45,7 @@ namespace enigma //TODO: This is from events.h: fix according to TODO pointed to
 extern std::string string(double val);
 
 extern void ABORT_ON_FATAL_ERRORS();
-int show_error(std::string errortext,double fatal)
+int show_error(std::string errortext,int fatal)
 {
   if (MessageBox(NULL,("Error in "+enigma::evname(enigma::event_current) + " event for object number " + string(enigma::getID())+": \r\n"+errortext).c_str(),"Error",MB_ABORTRETRYIGNORE | MB_ICONERROR)==IDABORT)
     exit(0);
